+++ boundlexx/api/v2/views/item.py
@@ -46,7 +46,11 @@ class ItemViewSet(
     queryset = (
         Item.objects.filter(active=True)
         .select_related("item_subtitle", "list_type", "description")
-        .prefetch_related("localizedname_set", "item_subtitle__localizedname_set")
+        .prefetch_related(
+            "localizedname_set",
+            "item_subtitle__localizedname_set",
+            "list_type__strings",
+        )
     )
     serializer_class = SimpleItemSerializer
     detail_serializer_class = ItemSerializer